working version 1

+++ join.js
@@ -1,33 +1,41 @@
 const path = require("path")
 const fs = require("fs")
 const util = require("util")
-const linkPreviewGenerator = require("link-preview-generator");
-//const xml2js = require('xml2js');
-//const https = require('https');
 const Parser = require('rss-parser');
+const seeLink = require('see-link');
 
-const rssurl = "https://rss.pimsnel.com/i/?a=rss&get=s&hours=16800";
-
+const rssurl = "https://rss.pimsnel.com/i/?a=rss&get=s&hours=168000000000000";
 const directoryPath = path.join(__dirname, "telegram");
-let jsonData = [];
+const filenameCache = './cache/previews-uid.json';
+
+async function exists (path) {
+  try {
+    await fs.accessSync(path)
+    return true
+  } catch {
+    return false
+  }
+}
 
 async function showPreview(link){
-  const previewData = await linkPreviewGenerator( link, [], 
-    "facebookexternalhit/1.1 (+http://www.facebook.com/externalhit_uatext.php)",
-    "chromium"
-  );
+  const previewData = await seeLink(link, {executablePath: 'chromium'});
   return previewData;
 }
 
 async function cachePreview(link, filename, type, source ){
 
+  if(await exists(filename)){
+    console.log("already in cache: "+ link);
+    return true;
+  }
+
   previewData = await showPreview(link);
-  console.log(previewData);
   previewData.source = source;
   previewData.type = type;
   const data = JSON.stringify(previewData);
   fs.writeFileSync(filename, data);
-  console.log("wrote"+ filename);
+  console.log("wrote preview of " + link + " to " + filename);
+  return true;
 }
 
 async function walkTelegramFiles(){
@@ -42,43 +50,22 @@ async function walkTelegramFiles(){
           var filearr = basefile.split("_");
 
           let itemArr = {};
-          //itemArr.permid = file;
-          //itemArr.date = filearr[0];
           itemArr.type = filearr[2];
-          //itemArr.source = "telegram";
-
           if(itemArr.type == "link"){
-            const link = fs.readFileSync(path.join(directoryPath, file),{encoding:'utf8', flag:'r'});
 
-            //itemArr.url = link;
-            //itemArr.preview = await showPreview(link);
+            const link = fs.readFileSync(path.join(directoryPath, file),{encoding:'utf8', flag:'r'});
             var filename = 'previews/'+filearr[0]+'_'+ filearr[1]+'_link.json';
-            cachePreview(await showPreview(link),filename,"link", "telegram");
-            //jsonData.push(itemArr);
+            cachePreview(link,filename,"link", "telegram");
           }
           else if(itemArr.type == "image"){
-            //itemArr.image = file;
             //jsonData.push(itemArr);
           }
         }
       };
-
-      // convert JSON object to string
-      /*
-      const data = JSON.stringify(jsonData);
-
-      // write JSON string to a file
-      fs.writeFile('out/telegram.json', data, (err) => {
-         if (err) {
-            throw err;
-         }
-         console.log("JSON data is saved.");
-      });
-      */
-
-
     }
   })
+
+  return true
 }
 
 async function walkRSS(){
@@ -92,20 +79,26 @@ async function walkRSS(){
 
     let feed = await parser.parseURL(rssurl);
 
+    //let i=0;
     for(const item of feed.items){
-      console.log(item.link);
+      /*
+      i++;
+      if(i > 5){
+        break;
+      }
+      */
       var pubDate = new Date(item.pubDate).toISOString().substring(0,10);
       var filename = 'previews/'+pubDate+'_'+item.guid+'_link.json';
       await cachePreview(item.link,filename, "link", "rss");
     };
 
   })();
+
+  return true;
 }
 
-//walkTelegramFiles();
+//////// MAIN
+
+walkTelegramFiles();
 walkRSS();
-
-
-
-
